Resolved conflicts manually and moved files to correct locations
