+++ client/src/dev-components/container-card/container-card.js
@@ -6,11 +6,10 @@ import InlineSVG from 'svg-inline-react'
 import classNames from 'classnames/bind'
 import apiCall from '../../api-call'
 // Styles
-import '../scss/atoms/_toast.scss'
-import styles from '../scss/views/dashboard-view.scss'
-import SlackIcon from '../assets/svg/baseline-slack-24px.svg'
-import { shareSlack } from '../globals'
+import SlackIcon from '../../assets/svg/baseline-slack-24px.svg'
+import { shareSlack } from '../../globals'
 import styles from './container-card.scss'
+import '../../scss/atoms/_toast.scss'
 // Icons
 import CloseIcon from '../../assets/svg/baseline-close-24px.svg'
 import StartIcon from '../../assets/svg/baseline-play_arrow-24px.svg'
