+++ client/src/dev-components/dashboard-view/dashboard-view.scss
@@ -1,7 +1,32 @@
 @import '../../scss/compulsory';
 
 @import '../../scss/atoms/button';
-@import '../../scss/atoms/button-container';
+@import '../../scss/atoms/tooltip';
+
+// Placeholders
+
+%accent {
+  box-shadow: 0 2px 5px $palette-shadow-default;
+  background: $palette-accent;
+  color: $palette-lighter;
+}
+
+%anchor-button-accent {
+  @extend %anchor-button;
+  @extend %accent;
+}
+
+.slack-share i {
+  $dimension: 20px;
+  height: $dimension;
+  width: $dimension;
+
+  // Visually centered vertically
+  &svg {
+    position: relative;
+    top: -1px;
+  }
+}
 
 .top-section {
   display: flex;
@@ -10,6 +35,11 @@
 
   padding: $spacer * 3;
   box-shadow: 0 0 10px $palette-shadow-light;
+
+  .top-section-buttons {
+    display: flex;
+    align-items: center;
+  }
 }
 
 .create-button {
@@ -35,7 +65,61 @@
   }
 }
 
+.instance-card {
+  flex: 1;
+  padding: $spacer;
+  margin: $spacer;
+  border-radius: $border-radius-default;
+  box-shadow: 0 0 10px $palette-shadow-default;
+
+  &:not(:first-of-type) {
+    margin-top: 0;
+  }
+
+  .header-container {
+    display: flex;
+    justify-content: space-between;
+    flex-flow: row-reverse wrap;
+
+    h5 {
+      text-transform: capitalize;
+      margin-right: auto;
+    }
+
+    .header-button-container {
+      display: flex;
+      justify-content: space-between;
+
+      button {
+        @extend %icon-button;
+      }
+    }
+  }
+}
+
 .button-container {
-  @extend %button-container;
+  display: flex;
+  justify-content: space-between;
+  margin-top: $spacer;
+
+  button {
+    @extend %icon-button;
+  }
+
+  a {
+    @extend %anchor-button;
+
+    &.accent {
+      @extend %accent;
+    }
+  }
+
+  & > button,  & > a {
+    margin-left: $spacer;
+
+    &:first-child {
+      margin-left: 0;
+    }
+  }
 }
