+++ client/src/dev-components/sign-in-slack/sign-in-slack.scss
@@ -1,5 +1,6 @@
-@import '../scss/atoms-organisms/button';
-@import '../scss/compulsory';
+@import '../../scss/compulsory';
+
+@import '../../scss/atoms/button';
 
 .slack-sign-button {
   @include reset-button;
