--- client/src/scss/views/dashboard-view.scss
@@ -1,124 +1,0 @@
-@import '../../scss/compulsory';
-
-@import '../atoms-organisms/button';
-@import '../atoms-organisms/tooltip';
-
-// Placeholders
-
-%accent {
-  box-shadow: 0 2px 5px $palette-shadow-default;
-  background: $palette-accent;
-  color: $palette-lighter;
-}
-
-%anchor-button-accent {
-  @extend %anchor-button;
-  @extend %accent;
-}
-
-.slack-share i {
-  $dimension: 20px;
-  height: $dimension;
-  width: $dimension;
-
-  // Visually centered vertically
-  &svg {
-    position: relative;
-    top: -1px;
-  }
-}
-
-.top-section {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  padding: $spacer * 3;
-  box-shadow: 0 0 10px $palette-shadow-light;
-
-  .top-section-buttons {
-    display: flex;
-    align-items: center;
-  }
-}
-
-.create-button {
-  @extend %anchor-button-accent;
-  align-self: flex-end;
-}
-
-.instances-container {
-  display: flex;
-  flex-flow: column;
-  justify-content: center;
-  align-items: center;
-  padding: $spacer * 2;
-
-  h2 {
-    margin-top: $spacer * 3;
-    margin-bottom: $spacer * 2;
-  }
-
-  .reverse {
-    display: flex;
-    flex-direction: column-reverse;
-  }
-}
-
-.instance-card {
-  flex: 1;
-  padding: $spacer;
-  margin: $spacer;
-  border-radius: $border-radius-default;
-  box-shadow: 0 0 10px $palette-shadow-default;
-
-  &:not(:first-of-type) {
-    margin-top: 0;
-  }
-
-  .header-container {
-    display: flex;
-    justify-content: space-between;
-    flex-flow: row-reverse wrap;
-
-    h5 {
-      text-transform: capitalize;
-      margin-right: auto;
-    }
-
-    .header-button-container {
-      display: flex;
-      justify-content: space-between;
-
-      button {
-        @extend %icon-button;
-      }
-    }
-  }
-}
-
-.button-container {
-  display: flex;
-  justify-content: space-between;
-  margin-top: $spacer;
-
-  button {
-    @extend %icon-button;
-  }
-
-  a {
-    @extend %anchor-button;
-
-    &.accent {
-      @extend %accent;
-    }
-  }
-
-  & > button,  & > a {
-    margin-left: $spacer;
-
-    &:first-child {
-      margin-left: 0;
-    }
-  }
-}
